Decouple api commands from build #18
