+++ cmd/hub/cmd/vars.go
@@ -12,21 +12,13 @@ const (
 	envVarNameState             = "HUB_STATE"
 	envVarNameAwsRegion         = "HUB_AWS_REGION"
 	envVarNameComponentsBaseDir = "HUB_COMPONENTS_BASEDIR"
-	envVarNameHubApi            = "HUB_API"
-	envVarNameDerefSecrets      = "HUB_API_DEREF_SECRETS"
-	SuperHubIo                  = ".superhub.io"
-
-	mdpre = "```"
 )
 
 var (
-	supportedClouds            = []string{"aws", "azure", "gcp"}
-	supportedCloudAccountKinds = []string{"aws", "azure", "gcp"}
+	supportedClouds = []string{"aws", "azure", "gcp"}
 )
 
 var (
-	environmentSelector   string
-	templateSelector      string
 	componentName         string
 	componentsBaseDir     string
 	elaborateManifest     string
@@ -36,11 +28,4 @@ var (
 	dryRun                bool
 	osEnvironmentMode     string
 	outputFiles           string
-	waitAndTailDeployLogs bool
-	showSecrets           bool
-	showLogs              bool
-	jsonFormat            bool
-	patchReplace          bool
-	patchRaw              bool
-	createRaw             bool
 )
